+++ packages/card/src/assets/basic/princess.json
