+++ packages/card/src/assets/far-eastern-border/princess.json
