+++ packages/card/src/far-eastern-border/princess.ts
@@ -1,3 +1,3 @@
-export { default as princesses } from "../../cards/far-eastern-border/princess.json" with {
+export { default as princesses } from "../assets/far-eastern-border/princess.json" with {
   type: "json",
 };
